refactor(sse): extrai a família GLM de services/usage.ts
Continua a decomposição do god-file services/usage.ts (empilhado sobre a fatia
MiniMax). Move a família GLM (Zhipu) completa para services/usage/glm.ts (181L):
naming de quota token/janela (getGlmTokenQuotaName/getGlmQuotaDisplayName/
getGlmQuotaLabel), ordenação (orderGlmQuotas + GLM_QUOTA_ORDER), math de
remaining mensal (glmMonthlyRemainingPercentage) e o fetcher getGlmUsage.

- Importa só os leaves irmãos scalars (toNumber/toRecord/toTitleCase/toPercentage)
  e quota (UsageQuota) + o config getGlmQuotaUrl — sem acoplamento com o host.
- usage.ts reimporta getGlmUsage (dispatcher) e re-exporta glmMonthlyRemainingPercentage
  (o teste glm-coding-plan-monthly importa de services/usage). shouldDisplayGitHubQuota
  (intercalado, não-GLM) preservado.

Behavior-preserving (byte-idêntico). usage.ts 2817 → 2670.

Validação: typecheck:core, eslint 0/0, check:cycles, check:file-size,
check:public-creds OK, glm-coding-plan-monthly (4/4), usage-utils (47/47),
usage-service-hardening (24/24) + caracterização nova (3/3, surface + identidade
do re-export + math de remaining mensal).
